requirements.txt
langchain
python-dotenv
ipykernel
langchain-google-genai
langchain_community
langchain_core
langchain_groq
langchain_chromadb
streamlit
google-generativeai
pymupdf
arxiv
wikipedia
mysql-connector-python
SQLALchemy
validators==0.28.1
youtube_transcript_api
unstructured
numexpr